+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ProofTree.expand_tree()` is a method the can be used to expand a proof tree
+  using a strategy pack. This can be used to expand strategy verified
+  combinatorial classes, to give a single tree.
+
+### Changed
+- The `CombinatorialSpecification.auto_seach()` only uses keyword arguments.
 
 ## [0.4.0] - 2020-02-20
 ### Fixed